Initial commit
Use this script to launch multiple iButtons with the same
starting time and sampling interval. It is meant to speed up the
process by filling in all of the options for the iButton
automatically instead of forcing you to manually enter them as is
the case with the Java-based iButton programs.