Modified skip-number routine
Added additional options when user elects to skip the next
file number in the sequence. The script now shows the user the
next file number and gives them an option to further increment or
decrement it. Once the new file number value is accepted, the script
pauses to let the user load the new ibutton into the reader.