Serial number search script, 1st commit
Use this script to generate a list of ibutton serial numbers
that you have successfully read (from the output csv files),
and then compare that list to the serial number of the ibutton(s)
in the reader currently. Meant to be used in conjunction with a
script like 'ibutton_download_increment_names.R'.