Major update - daylight savings handling
The original version of this script did not properly handle the
daylight savings time transition, despite my best intentions. It
used the current R timezone (inherited from the system) for
processing POSIX time values. If the system timezone was set to
a value that observes daylight savings time transitions, this
script would lose an hour of data when creating the output csv files.
I have now added a line at the start of the script to set the R
timezone to a timezone that does not observe daylight savings time
transitions. It must be set by the user for their local timezone
if you are not on the US Pacific coast time zone. See the notes
at the head of the script.

This really only matters if your iButtons were logging data during
a daylight savings transition (early March, early November in the
USA).